Create label-focused file label.py, and move classes for reading and writing strings into it
